ref(tiller): refactor tests into logical files
TestInstallRelease_WithChartAndDependencyNotes  -> tiller/release_install_test.go
TestInstallRelease_WrongTillerVersion -> tiller/release_install_test.go
TestInstallRelease_WithNotesRendered -> tiller/release_install_test.go
TestInstallRelease_TillerVersion -> tiller/release_install_test.go
TestInstallRelease_FailedHooks -> tiller/release_install_test.go
TestInstallRelease_ReuseName -> tiller/release_install_test.go
TestInstallRelease_WithNotes -> tiller/release_install_test.go
TestInstallRelease_NoHooks -> tiller/release_install_test.go
TestInstallRelease_DryRun -> tiller/release_install_test.go
TestInstallRelease -> tiller/release_install_test.go
